requirements.txt
requests
python-telegram-bot==22.6